feat(markdown): 支持文件预览公式和图表渲染
