+++ openspec/changes/add-file-markdown-math-preview/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-05-18
