+++ src/i18n/locales/en.part2.ts
@@ -541,6 +541,12 @@ const enPart2 = {
     annotationSubmit: "Annotate",
     markdownAnnotationToolbar: "Markdown preview annotation",
     markdownAnnotationHint: "Annotate source line numbers in preview without changing the file",
+    markdownFrontmatterLabel: "Metadata",
+    markdownMermaidTabList: "Mermaid block view",
+    markdownMermaidSource: "Source",
+    markdownMermaidRender: "Render",
+    markdownMermaidRendering: "Rendering diagram...",
+    markdownMermaidRenderFailed: "Render failed: {{message}}",
     annotationStartLine: "Start line",
     annotationEndLine: "End line",
     annotationSelectionToolbar: "Code selection annotation",
