+++ src/i18n/locales/zh.part2.ts
@@ -523,6 +523,12 @@ const zhPart2 = {
     annotationSubmit: "注释",
     markdownAnnotationToolbar: "Markdown 预览标注",
     markdownAnnotationHint: "预览模式按源文件行号标注，不会修改原文",
+    markdownFrontmatterLabel: "元数据",
+    markdownMermaidTabList: "Mermaid 图表视图",
+    markdownMermaidSource: "源码",
+    markdownMermaidRender: "渲染",
+    markdownMermaidRendering: "正在渲染图表...",
+    markdownMermaidRenderFailed: "渲染失败：{{message}}",
     annotationStartLine: "起始行",
     annotationEndLine: "结束行",
     annotationSelectionToolbar: "代码选区标注",
