+++ src/test/vitest.setup.ts
@@ -890,6 +890,12 @@ vi.mock("react-i18next", () => ({
         "threads.turnFailedToStartWithMessage": "会话启动失败：{{message}}",
         "threads.fusionTurnStalled": "融合回复未能接上，当前线程已回到可继续操作的状态。",
         "threads.fusionTurnStalledWithMessage": "融合回复未能接上：{{message}}",
+        "files.markdownMermaidSource": "Source",
+        "files.markdownMermaidRender": "Render",
+        "files.markdownMermaidTabList": "Mermaid block view",
+        "files.markdownMermaidRendering": "Rendering diagram...",
+        "files.markdownMermaidRenderFailed": "Render failed: {{message}}",
+        "files.markdownFrontmatterLabel": "Metadata",
         "threads.untitledThread": "未命名对话",
         "messages.middleStepsCollapsedHint": "已折叠 {{count}} 条中间步骤（实时中）",
         "workspace.homeHeroTitle": "构建任何东西",
